requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
cycler==0.10.0
Cython==0.28.5
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.0
Pillow==5.2.0
Pillow-PIL==0.1.dev0
pyparsing==2.2.0
pytesseract==0.2.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
